first commit of optional lzo
