+++ be/test/exec/CMakeLists.txt
@@ -39,7 +39,9 @@ ADD_BE_TEST(plain_text_line_reader_uncompressed_test)
 ADD_BE_TEST(plain_text_line_reader_gzip_test)
 ADD_BE_TEST(plain_text_line_reader_bzip_test)
 ADD_BE_TEST(plain_text_line_reader_lz4frame_test)
+if(DEFINED DORIS_WITH_LZO)
 ADD_BE_TEST(plain_text_line_reader_lzop_test)
+endif()
 ADD_BE_TEST(broker_reader_test)
 ADD_BE_TEST(broker_scanner_test)
 ADD_BE_TEST(broker_scan_node_test)
