+++ build.sh
@@ -51,15 +51,18 @@ Usage: $0 <options>
      --be               build Backend
      --fe               build Frontend
      --clean            clean and build target
-     --with-mysql       enable MySQL support
+     --with-mysql       enable MySQL support(default)
      --without-mysql    disable MySQL support
+     --with-lzo         enable LZO compress support(default)
+     --without-lzo      disable LZO compress  support
 
   Eg.
-    $0                          build Backend and Frontend without clean
-    $0 --be                     build Backend without clean
-    $0 --be --without-mysql     build Backend with MySQL disable
-    $0 --fe --clean             clean and build Frontend
-    $0 --fe --be --clean        clean and build both Frontend and Backend
+    $0                                      build Backend and Frontend without clean
+    $0 --be                                 build Backend without clean
+    $0 --be --without-mysql                 build Backend with MySQL disable
+    $0 --be --without-mysql --without-lzo   build Backend with both MySQL and LZO disable
+    $0 --fe --clean                         clean and build Frontend
+    $0 --fe --be --clean                    clean and build both Frontend and Backend
   "
   exit 1
 }
@@ -73,6 +76,8 @@ OPTS=$(getopt \
   -l 'clean' \
   -l 'with-mysql' \
   -l 'without-mysql' \
+  -l 'with-lzo' \
+  -l 'without-lzo' \
   -l 'help' \
   -- "$@")
 
@@ -87,6 +92,7 @@ BUILD_FE=
 CLEAN=
 RUN_UT=
 WITH_MYSQL=ON
+WITH_LZO=ON
 HELP=0
 if [ $# == 1 ] ; then
     # defuat
@@ -107,6 +113,8 @@ else
             --ut) RUN_UT=1   ; shift ;;
             --with-mysql) WITH_MYSQL=ON; shift ;;
             --without-mysql) WITH_MYSQL=OFF; shift ;;
+            --with-lzo) WITH_LZO=ON; shift ;;
+            --without-lzo) WITH_LZO=OFF; shift ;;
             -h) HELP=1; shift ;;
             --help) HELP=1; shift ;;
             --) shift ;  break ;;
@@ -126,10 +134,12 @@ if [ ${CLEAN} -eq 1 -a ${BUILD_BE} -eq 0 -a ${BUILD_FE} -eq 0 ]; then
 fi
 
 echo "Get params:
-    BUILD_BE -- $BUILD_BE
-    BUILD_FE -- $BUILD_FE
-    CLEAN    -- $CLEAN
-    RUN_UT   -- $RUN_UT
+    BUILD_BE    -- $BUILD_BE
+    BUILD_FE    -- $BUILD_FE
+    CLEAN       -- $CLEAN
+    RUN_UT      -- $RUN_UT
+    WITH_MYSQL  -- $WITH_MYSQL
+    WITH_LZO    -- $WITH_LZO
 "
 
 # Clean and build generated code
@@ -150,7 +160,7 @@ if [ ${BUILD_BE} -eq 1 ] ; then
     fi
     mkdir -p ${DORIS_HOME}/be/build/
     cd ${DORIS_HOME}/be/build/
-    cmake -DWITH_MYSQL=${WITH_MYSQL} ../
+    cmake -DWITH_MYSQL=${WITH_MYSQL} -DWITH_LZO=${WITH_LZO} ../
     make -j${PARALLEL}
     make install
     cd ${DORIS_HOME}
